+++ src/main/java/app/snapshot/qure/facilities/controller/FacilitiesController.java
@@ -28,9 +28,9 @@ public class FacilitiesController {
         this.geocodingService = geocodingService;
     }
 
-    private int mustManagerId() {
+    private Long mustManagerId() {
         // TODO: 로그인 붙이기 전 임시 값
-        return 1;
+        return 1L;
     }
 
 //    private Integer mustManagerId(HttpSession session) {
@@ -41,7 +41,7 @@ public class FacilitiesController {
     @GetMapping
     public String list(@RequestParam(value = "q", required = false) String q,
                        Model model, HttpSession session) {
-        int managerId = mustManagerId();
+        Long managerId = mustManagerId();
         List<FacilitiesDto> facilities = (q == null || q.isBlank())
                 ? facilitiesService.getFacilitiesListByManager(managerId)
                 : facilitiesService.searchFacilitiesByManager(managerId, q.trim());
@@ -61,7 +61,7 @@ public class FacilitiesController {
     public String create(@ModelAttribute FacilitiesDto dto,
                          RedirectAttributes ra,
                          HttpSession session) {
-        int managerId = mustManagerId();
+        Long managerId = mustManagerId();
 
         // 1) 주소 → 좌표 변환
         GeocodingService.LatLng ll = geocodingService.geocode(dto.getAddress());
@@ -76,7 +76,7 @@ public class FacilitiesController {
         }
 
         // 2) 시설과 QR 생성 + 좌표까지 한번에 저장
-        int facilityId = facilitiesService.createFacilityWithQr(dto, managerId);
+        Long facilityId = facilitiesService.createFacilityWithQr(dto, managerId);
 
         ra.addFlashAttribute("msg", "설비가 등록되었습니다.");
         return "redirect:/facilities";
@@ -89,8 +89,8 @@ public class FacilitiesController {
 
     // 등록 직후 QR 확인 페이지
     @GetMapping("/{facilityId}/qr")
-    public String showQr(@PathVariable int facilityId, Model model, RedirectAttributes ra) {
-        int managerId = mustManagerId();
+    public String showQr(@PathVariable Long facilityId, Model model, RedirectAttributes ra) {
+        Long managerId = mustManagerId();
 
         var facility = facilitiesService.findByfacilityIdAndManager(facilityId, managerId);
         if (facility == null) {
@@ -105,8 +105,8 @@ public class FacilitiesController {
 
     // 설비 수정
     @GetMapping("/{facilityId}/edit")
-    public String editForm(@PathVariable("facilityId") int facilityId, Model model, RedirectAttributes ra) {
-        int managerId = mustManagerId();
+    public String editForm(@PathVariable("facilityId") Long facilityId, Model model, RedirectAttributes ra) {
+        Long managerId = mustManagerId();
 
         FacilitiesDto facility = facilitiesService.findByfacilityIdAndManager(facilityId, managerId);
         if (facility == null) {
@@ -120,12 +120,12 @@ public class FacilitiesController {
 
     // 상세 (본인 소유만 접근)
     @GetMapping("/{facilityId}")
-    public String detail(@PathVariable int facilityId,
+    public String detail(@PathVariable Long facilityId,
                          @RequestParam(required = false) String start,
                          @RequestParam(required = false) String end,
                          Model model, RedirectAttributes ra, HttpSession session) {
 
-        int managerId = mustManagerId();
+        Long managerId = mustManagerId();
         FacilitiesDto facility = facilitiesService.findByfacilityIdAndManager(facilityId, managerId);
         if (facility == null) {
             ra.addFlashAttribute("msg", "접근 권한이 없거나 존재하지 않는 설비입니다.");
@@ -157,10 +157,10 @@ public class FacilitiesController {
 
     // 수정/삭제도 동일하게 managerId로 가드
     @PostMapping("/{id}")
-    public String edit(@PathVariable("id") int facilityId,
+    public String edit(@PathVariable("id") Long facilityId,
                        @ModelAttribute("facility") FacilitiesDto facility,
                        RedirectAttributes ra, HttpSession session) {
-        int managerId = mustManagerId();
+        Long managerId = mustManagerId();
         facility.setFacilityId(facilityId);
         int rows = facilitiesService.updateFacilitiesByManager(facility, managerId);
         ra.addFlashAttribute("msg", rows > 0 ? "설비가 수정되었습니다." : "수정 권한이 없거나 실패했습니다.");
@@ -168,8 +168,8 @@ public class FacilitiesController {
     }
 
     @PostMapping("/{facilityId}/delete")
-    public String delete(@PathVariable int facilityId, RedirectAttributes ra, HttpSession session) {
-        int managerId = mustManagerId();
+    public String delete(@PathVariable Long facilityId, RedirectAttributes ra, HttpSession session) {
+        Long managerId = mustManagerId();
         int cnt = facilitiesService.deleteFacilitiesByManager(facilityId, managerId);
         ra.addFlashAttribute("msg", cnt > 0 ? "삭제되었습니다." : "삭제 권한이 없거나 데이터가 없습니다.");
         return "redirect:/facilities";
